コメント追加・整備

+++ src/application/service/task.ts
@@ -17,9 +17,10 @@ export class TaskService {
     dueDate: string;
     priority: Priority;
   }) {
+    // ユーザーの存在とユーザーの権限の確認
     const user = await this.userRepository.findById(req.userId);
     if (!user) throw new Error("User not found.");
-    // if (!user.isAdmin()) throw new Error("Only admins can add tasks.");
+    if (!User.isAdmin(user.role)) throw new Error("Only admins can add tasks.");
 
     const task = new Task(
       req.title,
@@ -28,16 +29,19 @@ export class TaskService {
       req.priority
     );
 
+    // DBに保存
     return await this.taskRepository.save(task);
   }
 
   // タスク一覧の取得
   async getAllTasks(): Promise<Task[] | null> {
+    // タスク一覧をDBから検索
     return await this.taskRepository.findAllTasks();
   }
 
   // タスクの取得
   async getTaskById(taskId: string): Promise<Task | null> {
+    // タスクをDBから検索
     return await this.taskRepository.findById(taskId);
   }
 
@@ -50,26 +54,29 @@ export class TaskService {
     dueDate?: string;
     priority?: Priority;
   }) {
+    // タスクの存在確認
     const task = await this.taskRepository.findById(req.taskId);
     if (!task) throw new Error("Task not found.");
+    // ユーザーの存在とユーザーの権限の確認
     const user = await this.userRepository.findById(req.userId);
     if (!user) throw new Error("User not found.");
+    if (!User.isAdmin(user.role)) throw new Error("Only admins can add tasks.");
 
-    // 各プロパティがundefinedでない場合にのみタスクを更新
     if (
       req.title !== undefined ||
       req.description !== undefined ||
       req.dueDate !== undefined ||
       req.priority !== undefined
     ) {
-      // タスクの更新
       const updateTask: TaskDetail = {
         title: req.title,
         description: req.description,
         dueDate: req.dueDate,
         priority: req.priority,
       };
+      // 編集内容に応じてタスクデータを更新
       task.updateTask(updateTask);
+
       // DBに保存
       return await this.taskRepository.update(task);
     } else {
@@ -79,13 +86,16 @@ export class TaskService {
 
   // タスクの削除
   async deleteTask(req: { userId: string; taskId: string }) {
+    // タスクの存在確認
     const task = await this.taskRepository.findById(req.taskId);
     if (!task) throw new Error("Task not found.");
+    // ユーザーの存在とユーザーの権限の確認
     const user = await this.userRepository.findById(req.userId);
     if (!user) throw new Error("User not found.");
-    // if (!user.isAdmin()) throw new Error("Only admins can delete tasks.");
+    if (!User.isAdmin(user.role))
+      throw new Error("Only admins can delete tasks.");
 
-    // DBに保存
+    // DBから削除
     return await this.taskRepository.delete(req.taskId);
   }
